Made advances tracker ulong

+++ SWSH_OWRNG_Generator.WinForms/MainWindow.cs
@@ -956,13 +956,13 @@ namespace SWSH_OWRNG_Generator.WinForms
             return (s0, s1);
         }
 
-        public static int GetAdvancesPassed(ulong prevs0, ulong prevs1, ulong news0, ulong news1)
+        public static ulong GetAdvancesPassed(ulong prevs0, ulong prevs1, ulong news0, ulong news1)
         {
             if (prevs0 == news0 && prevs1 == news1)
                 return 0;
 
             var rng = new Xoroshiro128Plus(prevs0, prevs1);
-            int i = 0;
+            ulong i = 0;
             for (; i < 20_000; i++) // 20,000 is an arbitrary number to prevent an infinite loop
             {
                 rng.Next();
